fix bugs

+++ src/main/java/com/zezame/lipayz/service/impl/UserServiceImpl.java
@@ -1,5 +1,6 @@
 package com.zezame.lipayz.service.impl;
 
+import com.zezame.lipayz.config.RabbitMQConfig;
 import com.zezame.lipayz.constant.Message;
 import com.zezame.lipayz.constant.RoleCode;
 import com.zezame.lipayz.dto.CommonResDTO;
@@ -11,13 +12,19 @@ import com.zezame.lipayz.exceptiohandler.exception.ConflictException;
 import com.zezame.lipayz.exceptiohandler.exception.DuplicateException;
 import com.zezame.lipayz.exceptiohandler.exception.NotFoundException;
 import com.zezame.lipayz.mapper.PageMapper;
-import com.zezame.lipayz.repo.RoleRepo;
-import com.zezame.lipayz.mapper.UserMapper;
+import com.zezame.lipayz.model.Role;
 import com.zezame.lipayz.model.User;
+import com.zezame.lipayz.pojo.ActivateCustomerEmailPojo;
+import com.zezame.lipayz.repo.PaymentGatewayAdminRepo;
+import com.zezame.lipayz.repo.RoleRepo;
+import com.zezame.lipayz.repo.TransactionRepo;
 import com.zezame.lipayz.repo.UserRepo;
 import com.zezame.lipayz.service.BaseService;
 import com.zezame.lipayz.service.UserService;
+import com.zezame.lipayz.util.EmailUtil;
 import lombok.RequiredArgsConstructor;
+import org.springframework.amqp.rabbit.annotation.RabbitListener;
+import org.springframework.amqp.rabbit.core.RabbitTemplate;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.security.core.userdetails.UserDetails;
@@ -32,9 +39,12 @@ import java.util.ArrayList;
 public class UserServiceImpl extends BaseService implements UserService {
     private final UserRepo userRepo;
     private final RoleRepo roleRepo;
-    private final UserMapper userMapper;
     private final PageMapper pageMapper;
     private final PasswordEncoder passwordEncoder;
+    private final RabbitTemplate rabbitTemplate;
+    private final EmailUtil emailUtil;
+    private final TransactionRepo transactionRepo;
+    private final PaymentGatewayAdminRepo paymentGatewayAdminRepo;
 
     @Override
     public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
@@ -53,7 +63,7 @@ public class UserServiceImpl extends BaseService implements UserService {
     @Override
     public PageRes<UserResDTO> getUsers(Pageable pageable) {
         Page<User> users = userRepo.findAll(pageable);
-        return pageMapper.toPageResponse(users, userMapper::mapToDto);
+        return pageMapper.toPageResponse(users, this::mapToDto);
     }
 
     @Override
@@ -61,11 +71,18 @@ public class UserServiceImpl extends BaseService implements UserService {
         var userId = parseUUID(id);
         var user = userRepo.findById(userId)
                 .orElseThrow(() -> new NotFoundException("User Is Not Found"));
-        return userMapper.mapToDto(user);
+        return mapToDto(user);
+    }
+
+    private UserResDTO mapToDto(User user) {
+        var dto = new UserResDTO(
+                user.getId(), user.getFullName(), user.getRole().getName(), user.getVersion());
+
+        return dto;
     }
 
     @Override
-    public CreateResDTO registerCustomer(CreateUserReqDTO request) {
+    public CreateResDTO registerUser(CreateUserReqDTO request) {
         if (userRepo.existsByEmail(request.getEmail())) {
             throw new DuplicateException("Email Is Not Available");
         }
@@ -73,6 +90,14 @@ public class UserServiceImpl extends BaseService implements UserService {
         var role = roleRepo.findByCode(RoleCode.CUST.name())
                 .orElseThrow(() -> new NotFoundException("Role Is Not Found"));
 
+        var customer = createCustomer(request, role);
+
+        var savedCustomer = userRepo.save(prepareRegister(customer));
+        sendEmail(savedCustomer.getEmail(), savedCustomer.getActivationCode());
+        return new CreateResDTO(savedCustomer.getId(), Message.CREATED.getDescription());
+    }
+
+    private User createCustomer(CreateUserReqDTO request, Role role) {
         var activationCode = generateRandomAlphaNumeric(6);
 
         var customer = new User();
@@ -83,13 +108,42 @@ public class UserServiceImpl extends BaseService implements UserService {
         customer.setIsActivated(false);
         customer.setActivationCode(activationCode);
 
-        var savedCustomer = userRepo.save(prepareCreate(customer));
-        return new CreateResDTO(savedCustomer.getId(), Message.CREATED.getDescription());
+        return customer;
+    }
+
+    private void sendEmail(String email, String activationCode) {
+        var link = "http://localhost:8080/users/activate?email=" + email + "&code=" + activationCode;
+        var emailPojo = new ActivateCustomerEmailPojo(email, link);
+        rabbitTemplate.convertAndSend(
+                RabbitMQConfig.EMAIL_EX_ACTIVATION,
+                RabbitMQConfig.EMAIL_ROUTING_KEY_ACTIVATION,
+                emailPojo);
+    }
+
+    @RabbitListener(queues = RabbitMQConfig.EMAIL_QUEUE_ACTIVATION)
+    public void receiveEmailNotifcationActivation(ActivateCustomerEmailPojo emailPojo) {
+        emailUtil.sendEmail(emailPojo.getCustomerEmail(),
+                "Activation Link",
+                "Visit This Link To Activate Your Account " + emailPojo.getLink());
     }
 
     @Override
-    public CommonResDTO deleteCustomer(String id) {
+    public CommonResDTO deleteUser(String id) {
         var customer = findCustomerById(id);
+
+        if (transactionRepo.existsByCustomer(customer)) {
+            throw new ConflictException("Customer Cannot Be Deleted, Because They Have Transaction History");
+        }
+
+        if (customer.getRole().getCode().equals(RoleCode.PGA.name())) {
+            var paymentGatewayAdmin = paymentGatewayAdminRepo.findByUser(customer);
+            if (transactionRepo.existsByUpdatedByEquals(paymentGatewayAdmin.getId())) {
+                throw new ConflictException("This Payment Gateway Admin Cannot Be Deleted, Because They Have Transactions History");
+            }
+
+            paymentGatewayAdminRepo.delete(paymentGatewayAdmin);
+        }
+
         userRepo.delete(customer);
         return new CommonResDTO(Message.DELETED.getDescription());
     }
@@ -99,12 +153,12 @@ public class UserServiceImpl extends BaseService implements UserService {
         var customer = userRepo.findCustomerToActivate(email, code, RoleCode.CUST.name())
                 .orElseThrow(() -> new NotFoundException("Customer Is Not Found"));
 
-        if (customer.getIsActivated() == true) {
+        if (customer.getIsActivated()) {
             throw new ConflictException("This Account Has Been Activated");
         }
 
         customer.setIsActivated(true);
-        userRepo.saveAndFlush(prepareUpdate(customer));
+        userRepo.saveAndFlush(prepareActivate(customer));
         return new CommonResDTO(Message.UPDATED.getDescription());
     }
